peakfit: Editors in separate file
